+++ src/p40_flowbase/helpers/__init__.py
@@ -1,5 +1,10 @@
 """Helper utilities for p40_flowbase."""
 
+from p40_flowbase.helpers.file_stats import (
+    count_files,
+    dir_size_bytes,
+    file_or_dir_size_bytes,
+)
 from p40_flowbase.helpers.json_extract import extract_json_from_response
 from p40_flowbase.helpers.paths import safe_path_component
 from p40_flowbase.helpers.sql_templates import (
@@ -11,7 +16,10 @@ from p40_flowbase.helpers.templates import render_prompt_template
 
 __all__ = [
     "arrow_schema_from_pydantic",
+    "count_files",
+    "dir_size_bytes",
     "extract_json_from_response",
+    "file_or_dir_size_bytes",
     "render_prompt_template",
     "render_sql_template",
     "safe_path_component",